+++ src/abstraction/i-timer.ts
@@ -1,0 +1,11 @@
+export interface ITimer {
+  id: string;
+  type: new (...args) => any;
+  interval: number;
+  target: Object;
+  method: Function;
+  typeName: string;
+  methodName: string;
+  envProperty: string;
+  handle?: any;
+}