+++ src/abstraction/index.ts
@@ -1,3 +1,4 @@
 export * from './i-field';
 export * from './i-schema';
+export * from './i-timer';
 export * from './type';