+++ src/confirmation/index.ts
@@ -1,5 +1,6 @@
 export * from './is-browser';
 export * from './is-empty';
 export * from './is-json';
+export * from './is-numeric';
 export * from './is-primitive';
 export * from './is-xml';