+++ src/globalization/timers.ts
@@ -1,0 +1,2 @@
+import { ITimer } from '../abstraction/i-timer';
+export const timers = new Array<ITimer>();